Prefer <Photo> component where applicable
