+++ app/src/components/ui/layout/ImageItem/ImageItem.tsx
@@ -1,16 +1,6 @@
 import React from "react";
 import { makeStyles } from "@material-ui/styles";
-
-type StyleProps = {
-  imageURL?: string;
-};
-
-function backgroundImage({ imageURL }: StyleProps) {
-  if (imageURL) {
-    return `url(${imageURL})`;
-  }
-  return "none";
-}
+import Photo from "../Photo";
 
 const useStyles = makeStyles({
   ImageItem: {
@@ -20,18 +10,8 @@ const useStyles = makeStyles({
   },
   photo: {
     background: "#c9c9c9",
-    backgroundImage,
-    backgroundPosition: "center",
-    backgroundSize: "cover",
-    overflow: "hidden",
     height: "85px",
     width: "85px",
-    "& > img": {
-      height: "100%",
-      objectFit: "cover",
-      objectPosition: "center",
-      width: "100%",
-    },
   },
   meta: {
     color: "#5a5a5a",
@@ -49,10 +29,12 @@ interface ImageItemProps {
 }
 
 const ImageItem: React.FC<ImageItemProps> = ({ imageURL, children }) => {
-  const classes = useStyles({ imageURL });
+  const classes = useStyles();
   return (
     <div className={classes.ImageItem}>
-      <div className={classes.photo} />
+      <div className={classes.photo}>
+        <Photo url={imageURL} />
+      </div>
       <div className={classes.meta}>{children}</div>
     </div>
   );
